flyttet ut kvp avlsuttet fra aktivitet service (#164)
