+++ src/test/java/no/nav/veilarbaktivitet/service/KafkaConsumerServiceTest.java
@@ -2,6 +2,7 @@ package no.nav.veilarbaktivitet.service;
 
 import no.nav.veilarbaktivitet.domain.Person;
 import no.nav.veilarbaktivitet.domain.kafka.OppfolgingAvsluttetKafkaDTO;
+import no.nav.veilarbaktivitet.kvp.KVPAvsluttetService;
 import no.nav.veilarbaktivitet.kvp.KvpAvsluttetKafkaDTO;
 import org.junit.Test;
 
@@ -17,7 +18,7 @@ public class KafkaConsumerServiceTest {
     @Test
     public void skal_behandleOppfolgingAvsluttet() {
         AktivitetService aktivitetService = mock(AktivitetService.class);
-        KafkaConsumerService consumerService = new KafkaConsumerService(aktivitetService);
+        KafkaConsumerService consumerService = new KafkaConsumerService(aktivitetService, null);
 
         ZonedDateTime sluttdato = ZonedDateTime.of(2020, 4, 5, 16, 17, 0, 0, ZoneId.systemDefault());
         Date expectedSluttdato = new Date(sluttdato.toInstant().toEpochMilli());
@@ -28,13 +29,14 @@ public class KafkaConsumerServiceTest {
 
         consumerService.behandleOppfolgingAvsluttet(oppfolgingAvsluttetDto);
 
-        verify(aktivitetService, times(1)).settAktiviteterTilHistoriske(eq(Person.aktorId("1234")), eq(expectedSluttdato));
+        verify(aktivitetService, times(1)).settAktiviteterTilHistoriske(Person.aktorId("1234"), expectedSluttdato);
     }
 
     @Test
     public void skal_behandleKvpAvsluttet() {
         AktivitetService aktivitetService = mock(AktivitetService.class);
-        KafkaConsumerService consumerService = new KafkaConsumerService(aktivitetService);
+        KVPAvsluttetService kvpAvsluttetService = mock(KVPAvsluttetService.class);
+        KafkaConsumerService consumerService = new KafkaConsumerService(aktivitetService, kvpAvsluttetService);
 
         ZonedDateTime sluttdato = ZonedDateTime.of(2020, 4, 5, 16, 17, 0, 0, ZoneId.systemDefault());
         Date expectedSluttdato = new Date(sluttdato.toInstant().toEpochMilli());
@@ -46,8 +48,8 @@ public class KafkaConsumerServiceTest {
 
         consumerService.behandleKvpAvsluttet(kvpAvsluttetDto);
 
-        verify(aktivitetService, times(1)).settAktiviteterInomKVPPeriodeTilAvbrutt(
-                eq(Person.aktorId("1234")), eq("begrunnelse"), eq(expectedSluttdato)
+        verify(kvpAvsluttetService, times(1)).settAktiviteterInomKVPPeriodeTilAvbrutt(
+                Person.aktorId("1234"), "begrunnelse", expectedSluttdato
         );
     }
